pyproject.toml
[project]
name = "tsura"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Andre Petersen", email = "info@tsura.org"},
]
dependencies = [
    "Flask>=3.0.2",
    "SQLAlchemy>=2.0.29",
    "psycopg2-binary>=2.9.9",
    "Flask-SQLAlchemy>=3.1.1",
    "python-dotenv>=1.0.1",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
package-type = "application"
